catalog retrieval done
