--- sql_utils/urls.txt
@@ -1,20 +1,0 @@
-2d53935dfd7d9382e1c6a1c87e317653901a6e68ab98d5fe3a1b51afc57a690b WIDE_AEGIS forced
-ef4a0ea0a0d472782e123c227a3405b9ec61946c3fa041e90270b9fcf347071e WIDE_GAMA09H forced
-17fd544ea48ea98e93cb7e8adedd77a6b393c9dfe0b5bc88f6fd1f3263a56de2 WIDE_GAMA15H forced
-28c249258a8948e8f553e89bb6d9472c5ada96f6875751965da137956444a6c0 WIDE_HECTOMAP forced
-a4f5aba387eb56b1676846f55b62904137d15c7ec5ec9638598043607adcaf75 WIDE_WIDE12H forced
-2e4e62a7b79f2ccf6c350745ba554d65685e237c747a7b470b2f7f6de554a8f6 WIDE_XMMLSS forced
-8bfb79751833fdcdcd1102c23160ac272a684bbf8af43d3cdf04841772567694 WIDE_VVDS_part1 forced
-a2c86a175975b4292e96696015facb14fa5bb4793aec1d05c3dbb9f8197b1744 WIDE_VVDS_part2 forced
-2c070da65cd433605b7130b8462be19b54995f9e0bd5571a2ff5c564043d82c4 DEEP_COSMOS forced
-98f1b8fdb0fa60c10c5202ed1899ae42939ca506052c67b0ffbfd7cd4ec5c997 DEEP_DEEP23 forced
-e389d871bd41c3a7435fb0e39f761f34cd380b385eab41db949a360445d38b60 DEEP_ELAISN1 forced
-7ff6206ab0856fcdb80faec19380f281cd78eb9576b9ce23ee70b4c330da3f4b DEEP_XMMLSS forced
-7d2f0e9f1f1b75968a36a6657fdf1db605e93ab5004cb6bf4d94007f50984b29 UDEEP_COSMOS forced
-5fdc1e911d32a741f237b53575f44fef8c2423949f5d78924808484885c70bcb UDEEP_SXDS forced
-effbec4e7a6abb8f95718e1524afdb760b95afb2b7d3a5e173f664fa18be124a COSMOS_WIDE_BEST forced
-598a5c2837c27f46d08554ae21e572aecb9b82c3b7b9aa53fcbac9d57dd97a3b COSMOS_WIDE_MEDIAN forced
-c54bfb0320e06df9b64471dd63eb00c272e11a6717415778959e37e7d2bbf2a6 COSMOS_WIDE_WORST forced
-14ced4102a3069516e25a175697f94479f34e82923c4f02b0c70e8b15302a534 WIDE frames
-6ac0fc1f74295b6c9ae468e2a4184081507dade733c080c57cb4b6a127a509d2 DEEP frames
-c2c238d284662f79bef7bb5b096520b4943612243ccbbdfe16288d67e1eeac94 UDEEP frames
