--- sql_utils/urls_strict.txt
@@ -1,8 +1,0 @@
-e4806b457d3a35d9e7d41b08adbf729dff05d98893ee97a4e8dc28f961923665 WIDE_AEGIS forced
-b69a06d1b1b46b6adda0e71f531705c2bff2d80a3a5f800a7d8548c526e9a534 WIDE_GAMA09H forced
-b3d692e2da9df9ba9b9783ad0f3f095c572e9f6073aad09d48861cea6ed3ba92 WIDE_GAMA15H forced
-dca113c44e8789a159a973b6c7db575c7603f99e5f9fd8e48d4fe232197324bc WIDE_HECTOMAP forced
-e33829d35304c49fca775c654a40d9c4471a18489248920a0da1902cc585c14e WIDE_WIDE12H forced
-56067dfff9599dd7491851d6b6aa3d583e59958f0c759822354f0e13ea9b7965 WIDE_XMMLSS forced
-3f0b6e78ed464c0afc6b9e19d1d5a272e5373d9a9c970cd4626fa261c069ae52 WIDE_VVDS_part1 forced
-1209e3b8579a3f1c31b10501d1a192fbeed925f5234d6e4b08899f62fdd26cfa WIDE_VVDS_part2 forced
